Moved log in page to a component to enable render via RouteProvider in App
