+++ src/App.css
@@ -89,6 +89,11 @@ body{
     align-items: center;
     background: transparent;
 }
+
+.InputCont input{
+    border: 0;
+}
+
 .deskImg{
     display:none;
   }
